fix: useMedia devices bug not giving devices when permissions are allowed
Render jitsi player in Networking room container using JitsiPlayerComponent
