+++ src/Components/EventSession/JitsiPlayerComponent.js
@@ -13,7 +13,7 @@ const JitsiPlayerComponent =({
   avatarUrl,
   displayName,
   sessionId,
-  containerId,
+  // containerId,
   domain,
   showJitsiLogo,
   subject,
@@ -27,7 +27,7 @@ const JitsiPlayerComponent =({
     avatarUrl,
     displayName,
     sessionId,
-    containerId,
+    containerId:"#conference-container",
     domain,
     showJitsiLogo,
     subject,
