+++ src/Containers/EventSession/ConferenceRoomContainer.js
@@ -204,7 +204,7 @@ export default () => {
                 avatarUrl={user.avatarUrl}
                 displayName= {user.firstName + " " + user.lastName}
                 sessionId={sessionId}
-                containerId= "#conference-container"
+                // containerId= "#conference-container"
                 domain={domain}
                 showJitsiLogo={showJitsiLogo}
                 subject={subject}
